Extract JekyllCommandHelper module
