+++ lib/site-builder.js
@@ -6,6 +6,7 @@ var fs = require('fs');
 var path = require('path');
 var buildLogger = require('./build-logger');
 var Options = require('./options');
+var JekyllCommandHelper = require('./jekyll-command-helper');
 var fileLockedOperation = require('file-locked-operation');
 var childProcess = require('child_process');
 
@@ -223,35 +224,8 @@ SiteBuilder.prototype.readOrWriteConfig = function(configExists) {
   });
 };
 
-function JekyllCommandHelper(builder) {
-  this.builder = builder;
-  this.jekyll = config.jekyll;
-  this.args = ['build', '--trace', '--destination'];
-
-  if (builder.usesBundler) {
-    this.jekyll = config.bundler;
-    this.args = ['exec', 'jekyll'].concat(this.args);
-  }
-}
-
-JekyllCommandHelper.prototype.spawn = function(destination, extraConfig) {
-  var configs = ['_config.yml'];
-  if (extraConfig) {
-    configs.push(extraConfig);
-  }
-  configs.push(this.builder.opts.pagesConfig);
-  configs = configs.join(',');
-
-  if (this.builder.opts.branchInUrlPattern) {
-    destination = destination + '/' + this.builder.branch;
-  }
-
-  var args = this.args.concat(destination).concat('--config').concat(configs);
-  return this.builder.spawn(this.jekyll, args);
-};
-
 SiteBuilder.prototype.jekyllBuild = function() {
-  var helper = new JekyllCommandHelper(this);
+  var helper = new JekyllCommandHelper(this, config.jekyll, config.bundler);
   if (!this.hasInternalConfig) {
     return helper.spawn(this.buildDestination);
   }
